+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * :hammer: Activityx.getImplWithParent method marked static
 * :hammer: Color All method names that have started with make are changed to create
 * :hammer: Intent All method names that have started with make are changed to create
+* :fire: Remove Windowx isLandscape, isPortrait method
 
 ### v0.2.3
 * :hammer: Network getIpAddress rename to getLocalIpAddress